+++ src/dataflow/src/source/postgres.rs
@@ -24,6 +24,7 @@ use tokio_postgres::error::{DbError, Severity, SqlState};
 use tokio_postgres::replication::LogicalReplicationStream;
 use tokio_postgres::types::PgLsn;
 use tokio_postgres::SimpleQueryMessage;
+use tracing::{info, warn};
 
 use crate::source::{SimpleSource, SourceError, SourceTransaction, Timestamper};
 use mz_dataflow_types::{sources::PostgresSourceConnector, SourceErrorDetails};
@@ -444,7 +445,7 @@ impl SimpleSource for PostgresSourceReader {
                 let mut writer = tokio::io::BufWriter::new(file);
                 match self.produce_snapshot(&mut snapshot_tx, &mut writer).await {
                     Ok(_) => {
-                        tracing::info!(
+                        info!(
                             "replication snapshot for source {} succeeded",
                             &self.source_name
                         );
@@ -455,10 +456,9 @@ impl SimpleSource for PostgresSourceReader {
                             source_name: self.source_name.clone(),
                             error: SourceErrorDetails::Initialization(e.to_string()),
                         })?;
-                        tracing::warn!(
+                        warn!(
                             "replication snapshot for source {} failed, retrying: {}",
-                            &self.source_name,
-                            e
+                            &self.source_name, e
                         );
                         let reader = BufReader::new(writer.into_inner().into_std().await);
                         self.revert_snapshot(&mut snapshot_tx, reader)
@@ -484,10 +484,9 @@ impl SimpleSource for PostgresSourceReader {
         loop {
             match self.produce_replication(timestamper).await {
                 Err(ReplicationError::Recoverable(e)) => {
-                    tracing::warn!(
+                    warn!(
                         "replication for source {} interrupted, retrying: {}",
-                        &self.source_name,
-                        e
+                        &self.source_name, e
                     )
                 }
                 Err(ReplicationError::Fatal(e)) => {
@@ -501,7 +500,7 @@ impl SimpleSource for PostgresSourceReader {
 
             // TODO(petrosagg): implement exponential back-off
             tokio::time::sleep(Duration::from_secs(3)).await;
-            tracing::info!("resuming replication for source {}", &self.source_name);
+            info!("resuming replication for source {}", &self.source_name);
         }
     }
 }